cmds: pjsip: Change mmap_device_memory() parameter type from uint64_t to uintptr_t

+++ project/pjsip/cmds/sip-nuklear/nuklear_main.c
@@ -182,9 +182,9 @@ int demo_nk_main(const char *input_dev_path) {
 	screensize = fb_info->var.xres * fb_info->var.yres * bpp;
 
 	/* Map the device to memory */
-	fbp = (uint8_t *) mmap_device_memory((void *) fb_info->screen_base,
-			screensize, PROT_READ | PROT_WRITE, MAP_SHARED,
-			(uint64_t) ((uintptr_t) fb_info->screen_base));
+	fbp = (uint8_t *)mmap_device_memory((void *)fb_info->screen_base,
+	    screensize, PROT_READ | PROT_WRITE, MAP_SHARED,
+	    (uintptr_t)fb_info->screen_base);
 	if ((int) fbp == -1) {
 		perror("Error: failed to map framebuffer device to memory");
 		exit(4);
